cfg: set Vth_on to 2.2v
2.5v is too high: we boot up to that voltage only once,
because max Vbqout drops as the supercaps discharge.

Vbqout oscillates between an (ever decreasing) max
and an ever decreasing min. Note that min is not
necessarily below 1.8v, but eventually it gets there.
The oscillation seems to be because the BQ is not
generating current into the output continuously,
but has a hysterisis.

With 2.2v we get intermittent behavior when Vbqout's
min crosses 1.8v.